:recycle: Add custom hook
#205 handleUploadImg 함수를 커스텀 훅으로 생성하여 적용함
